feat:(LR-07): Added footer
