+++ src/App.tsx
@@ -1,3 +1,4 @@
+import { Footer } from '@components/Footer/footer.component'
 import { Header } from '@components/Header/header.component'
 import { Tabs } from '@ui/Tabs/tabs.component'
 
@@ -6,6 +7,7 @@ function App() {
     <>
       <Header />
       <Tabs />
+      <Footer />
     </>
   )
 }
